Add mdg retail code

+++ woocommerce_loyalty_points/models/sale_order.py
@@ -5,6 +5,6 @@ class sale_order(osv.osv):
     _inherit = 'sale.order'    
     
     _columns = {
-         'getting_point': fields.integer('Getting Point'),
-         'redeem_point': fields.integer('Redeem Point'),
+         'getting_point': fields.float('Getting Point'),
+         'redeem_point': fields.float('Redeem Point'),
     }
